Move getHistoricalSummaries to lodestar namespace
